+++ MS3KOS/src/msGlobals.h
@@ -21,7 +21,7 @@ class MagicShifterGlobals {
 	int ggAccelCount[3];		// Stores the 12-bit signed value
 	long ggAccelTime;
 	unsigned long ggLFrameTime;
-	long ggCurrentMode = DEFAULT_FACTORY_MODE;	// POV mode
+	uint32_t ggCurrentMode = DEFAULT_FACTORY_MODE;	// POV mode
 	int ggBtn1State = 0;
 	long ggCurrentMicros = 0;
 	unsigned long ggCurrentMillis = 0;
